Make Git exclude files when generating an archive
GitHub generates two archives (a `.zip` and `.tar.gz`) for every
tag/release we create. This is useful as we can provide the links
to those archives to users to download a specific version of our
project.

However, these archives will contain all files the project had
when the tag/release was made. This includes files that aren't of
any use to the users and are most likely going to be deleted
(e.g.: CONTRIBUTING.md).

This commit whitelists the files that aren't useful for users and
informs Git (and thus, GitHub) not to include them when generating
an archive.

Ref: http://git-scm.com/book/en/Customizing-Git-Git-Attributes#Exporting-Your-Repository