Drop _1 for logos for which the version is unknown
For some of the logos from the archive the exact version of the
browser that used them is unknown, therefore, to distinguish
between logos belonging to the same browser and for which the
browser version is unknown, `_1`, `_2`, and so on are added to
their file names.

This commit removes the `_1` for all cases in which a particular
browser has only one logo and the browser version is unknown.
This is done namely because the naming convention described above
were used inconsistently (some had the `_1` and some didn't), and
that created some confusion.